Share reindexer between elasticsearch and pgsearch
