+++ cmd/authgear/elasticsearch/reindex.go
@@ -20,6 +20,7 @@ import (
 	libes "github.com/authgear/authgear-server/pkg/lib/elasticsearch"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db"
 	"github.com/authgear/authgear-server/pkg/lib/infra/db/appdb"
+	"github.com/authgear/authgear-server/pkg/lib/search/reindex"
 	"github.com/authgear/authgear-server/pkg/util/graphqlutil"
 )
 
@@ -78,7 +79,7 @@ func (q *Reindexer) QueryPage(after model.PageCursor, first uint64) ([]Item, err
 		// To access standard attributes publicly, it should go through
 		// DeriveStandardAttributes func.
 		rawStandardAttributes := u.StandardAttributes
-		raw := &model.ElasticsearchUserRaw{
+		raw := &model.SearchUserRaw{
 			ID:                 u.ID,
 			AppID:              string(q.AppID),
 			CreatedAt:          u.CreatedAt,
@@ -175,8 +176,8 @@ func (q *Reindexer) reindex(ctx context.Context, bulkIndexer esutil.BulkIndexer)
 
 		// Process the items
 		for _, item := range items {
-			user := item.Value.(*model.ElasticsearchUserRaw)
-			source := libes.RawToSource(user)
+			user := item.Value.(*model.SearchUserRaw)
+			source := reindex.RawToSource(user)
 			allUserIDs[user.ID] = struct{}{}
 
 			var bodyBytes []byte
