+++ pkg/api/model/user.go
@@ -43,7 +43,7 @@ type UserRef struct {
 	Meta
 }
 
-type ElasticsearchUserRaw struct {
+type SearchUserRaw struct {
 	ID                 string
 	AppID              string
 	CreatedAt          time.Time
@@ -57,7 +57,7 @@ type ElasticsearchUserRaw struct {
 	StandardAttributes map[string]interface{}
 }
 
-type ElasticsearchUserSource struct {
+type SearchUserSource struct {
 	ID          string     `json:"id,omitempty"`
 	AppID       string     `json:"app_id,omitempty"`
 	CreatedAt   time.Time  `json:"created_at,omitempty"`
