+++ pkg/lib/search/reindex/deps.go
@@ -1,0 +1,9 @@
+package reindex
+
+import (
+	"github.com/google/wire"
+)
+
+var DependencySet = wire.NewSet(
+	wire.Struct(new(Reindexer), "*"),
+)
